refactor: split OPDS tests.rs and fix import order (#1829)
Splits opds/tests.rs (1657 lines) into tests/{mod,basic_auth,json_catalog,all_series_shelves}.rs
(all under 800 lines), mirroring the db/src/metadata_lookup/tests/ sibling-file pattern, and
reorders the crate::/super:: import blocks in all.rs, json_all.rs, series.rs, shelves.rs, and
json_shelves.rs to match delegate.rs's convention. Pure reorganization — all 63 test cases
carried over unchanged.

Closes #1829
